Extract wallet-rleated tests to separate file
