+++ solidity/test/bridge/Bridge.test.ts
@@ -14,7 +14,6 @@ import type {
   Bridge,
   BridgeStub,
   BridgeStub__factory,
-  Bridge__factory,
   TestRelay,
   TestRelay__factory,
   IWalletRegistry,
@@ -228,194 +227,6 @@ describe("Bridge", () => {
             .withArgs(vault, false)
         })
       })
-    })
-  })
-
-  describe("requestNewWallet", () => {
-    before(async () => {
-      await createSnapshot()
-    })
-
-    after(async () => {
-      walletRegistry.requestNewWallet.reset()
-
-      await restoreSnapshot()
-    })
-
-    // TODO: Implement wallet creation rules
-
-    context("when called by a third party", async () => {
-      it("should call ECDSA Wallet Registry's requestNewWallet function", async () => {
-        await bridge.connect(thirdParty).requestNewWallet(NO_MAIN_UTXO)
-
-        await expect(walletRegistry.requestNewWallet).to.have.been.calledOnce
-      })
-    })
-  })
-
-  describe("__ecdsaWalletCreatedCallback", () => {
-    context("when called by a third party", async () => {
-      it("should revert", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .__ecdsaWalletCreatedCallback(
-              ecdsaWalletTestData.walletID,
-              ecdsaWalletTestData.publicKeyX,
-              ecdsaWalletTestData.publicKeyY
-            )
-        ).to.be.revertedWith("Caller is not the ECDSA Wallet Registry")
-      })
-    })
-
-    context("when called by the ECDSA Wallet Registry", async () => {
-      context("when called with a valid ECDSA Wallet details", async () => {
-        let tx: ContractTransaction
-
-        before(async () => {
-          await createSnapshot()
-
-          tx = await bridge
-            .connect(walletRegistry.wallet)
-            .__ecdsaWalletCreatedCallback(
-              ecdsaWalletTestData.walletID,
-              ecdsaWalletTestData.publicKeyX,
-              ecdsaWalletTestData.publicKeyY
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should register ECDSA wallet reference", async () => {
-          await expect(
-            (
-              await bridge.getRegisteredWallet(
-                ecdsaWalletTestData.pubKeyHash160
-              )
-            ).ecdsaWalletID
-          ).equals(ecdsaWalletTestData.walletID)
-        })
-
-        it("should transition Wallet to Live state", async () => {
-          await expect(
-            (
-              await bridge.getRegisteredWallet(
-                ecdsaWalletTestData.pubKeyHash160
-              )
-            ).state
-          ).equals(walletState.Live)
-        })
-
-        it("should emit NewWalletRegistered event", async () => {
-          await expect(tx)
-            .to.emit(bridge, "NewWalletRegistered")
-            .withArgs(
-              ecdsaWalletTestData.walletID,
-              ecdsaWalletTestData.pubKeyHash160
-            )
-        })
-      })
-
-      context(
-        "when called with the ECDSA Wallet already registered",
-        async () => {
-          before(async () => {
-            await createSnapshot()
-
-            await bridge
-              .connect(walletRegistry.wallet)
-              .__ecdsaWalletCreatedCallback(
-                ecdsaWalletTestData.walletID,
-                ecdsaWalletTestData.publicKeyX,
-                ecdsaWalletTestData.publicKeyY
-              )
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          const testData = [
-            {
-              testName: "with unique wallet ID and unique public key",
-              walletID: ethers.utils.randomBytes(32),
-              publicKeyX: ethers.utils.randomBytes(32),
-              publicKeyY: ethers.utils.randomBytes(32),
-              expectedError: undefined,
-            },
-            {
-              testName: "with duplicated wallet ID and unique public key",
-              walletID: ecdsaWalletTestData.walletID,
-              publicKeyX: ethers.utils.randomBytes(32),
-              publicKeyY: ethers.utils.randomBytes(32),
-              expectedError: undefined,
-            },
-            {
-              testName:
-                "with unique wallet ID, unique public key X and duplicated public key Y",
-              walletID: ethers.utils.randomBytes(32),
-              publicKeyX: ethers.utils.randomBytes(32),
-              publicKeyY: ecdsaWalletTestData.publicKeyY,
-              expectedError: undefined,
-            },
-            {
-              testName:
-                "with unique wallet ID, unique public key Y and duplicated public key X",
-              walletID: ethers.utils.randomBytes(32),
-              publicKeyX: ecdsaWalletTestData.publicKeyY,
-              publicKeyY: ethers.utils.randomBytes(32),
-              expectedError: undefined,
-            },
-            {
-              testName: "with unique wallet ID and duplicated public key",
-              walletID: ethers.utils.randomBytes(32),
-              publicKeyX: ecdsaWalletTestData.publicKeyX,
-              publicKeyY: ecdsaWalletTestData.publicKeyY,
-              expectedError: "ECDSA wallet has been already registered",
-            },
-            {
-              testName: "with duplicated wallet ID and duplicated public key",
-              walletID: ecdsaWalletTestData.walletID,
-              publicKeyX: ecdsaWalletTestData.publicKeyX,
-              publicKeyY: ecdsaWalletTestData.publicKeyY,
-              expectedError: "ECDSA wallet has been already registered",
-            },
-          ]
-
-          testData.forEach((test) => {
-            context(test.testName, async () => {
-              beforeEach(async () => {
-                await createSnapshot()
-              })
-
-              afterEach(async () => {
-                await restoreSnapshot()
-              })
-
-              it(
-                test.expectedError ? "should revert" : "should not revert",
-                async () => {
-                  const tx: Promise<ContractTransaction> = bridge
-                    .connect(walletRegistry.wallet)
-                    .__ecdsaWalletCreatedCallback(
-                      test.walletID,
-                      test.publicKeyX,
-                      test.publicKeyY
-                    )
-
-                  if (test.expectedError) {
-                    await expect(tx).to.be.revertedWith(test.expectedError)
-                  } else {
-                    await expect(tx).not.to.be.reverted
-                  }
-                }
-              )
-            })
-          })
-        }
-      )
     })
   })
